requirements.txt
blinker==1.7.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discum==1.4.1
filetype==1.2.0
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
ua-parser==0.18.0
urllib3==2.0.7
websocket-client==0.59.0
Werkzeug==3.0.1
